+++ Sources/DbConnectionExtensions.Entity.cs
@@ -8,6 +8,39 @@ using System.Data;
 public static partial class DbConnectionExtensions {
 
 	/// <summary>
+	/// Counts all entities.
+	/// </summary>
+	/// <typeparam name="T">The entity type.</typeparam>
+	/// <param name="connection">The connection to the data source.</param>
+	/// <param name="timeout">The wait time, in seconds, before terminating the attempt to execute the command and generating an error.</param>
+	/// <param name="transaction">The transaction within which the command executes.</param>
+	/// <param name="builder">An optional command builder used to build the SQL query to be executed.</param>
+	/// <returns><see langword="true"/> if the specified entity has been deleted, otherwise <see langword="false"/>.</returns>
+	public static int CountAll<T>(this IDbConnection connection, int timeout = 30, IDbTransaction? transaction = null, SqlCommandBuilder? builder = null) where T: new() {
+		var (command, parameters) = (builder ?? SqlCommandBuilder.Create(connection)).GetCountAllCommand<T>();
+		command.Timeout = timeout;
+		command.Transaction = transaction;
+		return ExecuteScalar<int>(connection, command, parameters);
+	}
+
+	/// <summary>
+	/// Counts all entities.
+	/// </summary>
+	/// <typeparam name="T">The entity type.</typeparam>
+	/// <param name="connection">The connection to the data source.</param>
+	/// <param name="timeout">The wait time, in seconds, before terminating the attempt to execute the command and generating an error.</param>
+	/// <param name="transaction">The transaction within which the command executes.</param>
+	/// <param name="builder">An optional command builder used to build the SQL query to be executed.</param>
+	/// <param name="cancellationToken">The token to cancel the operation.</param>
+	/// <returns><see langword="true"/> if the specified entity has been deleted, otherwise <see langword="false"/>.</returns>
+	public static async Task<int> CountAllAsync<T>(this IDbConnection connection, int timeout = 30, IDbTransaction? transaction = null, SqlCommandBuilder? builder = null, CancellationToken cancellationToken = default) where T: new() {
+		var (command, parameters) = (builder ?? SqlCommandBuilder.Create(connection)).GetCountAllCommand<T>();
+		command.Timeout = timeout;
+		command.Transaction = transaction;
+		return await ExecuteScalarAsync<int>(connection, command, parameters, cancellationToken);
+	}
+
+	/// <summary>
 	/// Deletes the specified entity.
 	/// </summary>
 	/// <typeparam name="T">The entity type.</typeparam>